Add Rainier regulators JSON config file
Add the JSON configuration file that controls how voltage regulators are
configured and monitored on Rainier systems.

This configuration file is used by the phosphor-regulators application
on the BMC.

The configuration file name is controlled by another BMC application
(EntityManager) and is fairly generic.  For this reason, I am putting
the JSON file in a 'regulators' subdirectory.  This will hopefully avoid
future name conflicts in case other Rainier JSON files are stored in
this repository.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>