Remove AVDD voltage setting from JSON config file
The AVDD, 3.3VA, and 3.3VB regulators do not have standby power on
Rainier.  As a result, they cannot be configured.  Configuration occurs
at the beginning of the boot before the power sequencer has turned on
main power.

The 3.3VA and 3.3VB regulators were not being configured in the current
JSON file, but the AVDD regulator was.  Removed the AVDD configuration.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>